Sprint 6 (#589 E4): split PdfGenerator rendering from data assembly
Extracts the HTML-rendering + placeholder/QR processing out of PdfGenerator into
a new instance-based PdfHtmlRenderer (generate_html, generate_default_html,
get_appointment_receipt_template, process_qrcode_placeholders,
get_qr_code_target_url, process_validation_url_placeholders,
parse_validation_url_params, and the generate_magic_link_qr impl). PdfGenerator
keeps data assembly (generate_pdf_data, enrich_submission_data,
generate_pdf_data_from_form, generate_appointment_pdf_data, resolve_effective_
schedule) and composes the renderer in its constructor.

The 🔴 cpf_rf_encrypted 3-tier legacy fallback stays untouched on PdfGenerator.
PdfGenerator::generate_magic_link_qr is preserved as a thin public-static
delegator (FQN/contract intact). PdfGenerator 901 → 501 lines.

Renderer carries @covers + a setUp preload so pcov attributes the delegated
coverage (verified 93/112). Generated PDF HTML is byte-identical.
Behavior-preserving — Pdf suite green (68 tests); PHPStan L8 + WPCS clean.
